// src/module/material-management/material-requests/utils/enums.ts
enum MATERIAL_REQUEST_ASSIGN_TO {
	FOREMAN = "FOREMAN",
	REQUESTING_EMPLOYEE = "REQUESTING_EMPLOYEE",
	WAREHOUSE_MANAGER = "WAREHOUSE_MANAGER",
	OFFICE_MANAGER = "OFFICE_MANAGER",
	PROCUREMENT_SPECIALIST = "PROCUREMENT_SPECIALIST",
	CUSTOM_ASSIGNMENT = "CUSTOM_ASSIGNMENT",
}

enum MATERIAL_REQUEST_FILTER_TYPE {
	JOB_FILTER = "jobFilter",
	JOB_NUMBER_FILTER = "jobNumberFilter",
	BUILDER_FILTER = "builderFilter",
	PROJECT_FILTER = "projectFilter",
	PHASE_FILTER = "phaseFilter",
	KEYWORD_FILTER = "keywordFilter",
	REQUEST_NUMBER_FILTER = "requestNumberFilter",
	MODEL_FILTER = "modelFilter",
	DEPARTMENT_FILTER = "departmentFilter",
	PART_CODE_FILTER = "partCodeFilter",
	PART_NAME_FILTER = "partNameFilter",
	REASON_FILTER = "reasonFilter",
	REQUESTED_BY_FILTER = "requestedByFilter",
}

enum MATERIAL_REQUEST_NOTE_MODE {
	UPDATE = "update",
	ADD = "add",
}

enum MATERIAL_REQUEST_VIEW_MODE {
	ALL = "all",
	MY_REQUESTS = "my-requests",
}

enum MATERIAL_REQUEST_AUDIT_EVENT_TYPE {
	SUBMITTED = "SUBMITTED",
	ASSIGNED = "ASSIGNED",
	REASSIGNED = "REASSIGNED",
	NOTE_ADDED = "NOTE_ADDED",
	APPROVED = "APPROVED",
	REJECTED = "REJECTED",
}

enum MATERIAL_REQUEST_TYPE {
	MATERIAL = "Material",
	MISSING_ITEM = "Missing Item",
}

enum REQUEST_TYPE_FILTER_VALUE {
	PULL_LIST_REQUEST = "pull_list_request",
	NOT_IN_PULL_LIST = "not_in_pull_list",
	UNKNOWN_ITEM = "unknown_item",
}

export {
	MATERIAL_REQUEST_ASSIGN_TO,
	MATERIAL_REQUEST_FILTER_TYPE,
	MATERIAL_REQUEST_NOTE_MODE,
	MATERIAL_REQUEST_VIEW_MODE,
	MATERIAL_REQUEST_AUDIT_EVENT_TYPE,
	MATERIAL_REQUEST_TYPE,
	REQUEST_TYPE_FILTER_VALUE,
};